requirements.txt
flask 
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
psycopg2
cloudinary
Flask-Mail